refactor: /seminar server action 적용
