+++ app/[locale]/community/seminar/SeminarContent.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import SeminarSearchBar from '@/app/[locale]/community/seminar/SearchBar';
-import SeminarRow from '@/app/[locale]/community/seminar/SeminarRow';
-import SeminarYear from '@/app/[locale]/community/seminar/SeminarYear';
+import SeminarRow from '@/app/[locale]/community/seminar/helper/SeminarRow';
+import SeminarSearchBar from '@/app/[locale]/community/seminar/helper/SeminarSearchBar';
+import SeminarYear from '@/app/[locale]/community/seminar/helper/SeminarYear';
 
 import LoginVisible from '@/components/common/LoginVisible';
 import Pagination from '@/components/common/Pagination';
@@ -13,13 +13,14 @@ import { SeminarList } from '@/types/seminar';
 import { useCustomSearchParams } from '@/utils/hooks/useCustomSearchParams';
 import useResponsive from '@/utils/hooks/useResponsive';
 
-import AdminFeatures from './AdminFeatures';
+import AdminFeatures from './helper/AdminFeatures';
 
-const postsCountPerPage = 10;
+const POSTS_COUNT_PER_PAGE = 10;
 
 export default function SeminarContent({ data: { searchList, total } }: { data: SeminarList }) {
   const { page, keyword, setSearchParams } = useCustomSearchParams();
   const { screenType } = useResponsive();
+  const pageLimit = screenType == 'desktop' ? 10 : 5;
 
   const setCurrentPage = (pageNum: number) => {
     setSearchParams({ purpose: 'navigation', pageNum });
@@ -30,28 +31,33 @@ export default function SeminarContent({ data: { searchList, total } }: { data: 
       <div className="flex flex-row items-center gap-6">
         <SeminarSearchBar keyword={keyword} setSearchParams={setSearchParams} />
       </div>
+
       <div className="mb-8 mt-10 flex flex-col border-b-[1px] border-neutral-200">
-        {searchList.length > 0 ? (
+        {searchList.length === 0 ? (
+          <NoSearchResult />
+        ) : (
           searchList.map((post, index) => (
             <div key={post.id}>
               {post.isYearLast && <SeminarYear index={index} startDate={post.startDate} />}
               <SeminarRow seminar={post} hideDivider={false} />
             </div>
           ))
-        ) : (
-          <p className="mx-2.5 mb-8 mt-6">검색 결과가 존재하지 않습니다.</p>
         )}
       </div>
+
       <Pagination
-        totalPostsCount={total ?? 0}
-        postsCountPerPage={postsCountPerPage}
-        pageLimit={screenType == 'desktop' ? 10 : 5}
+        totalPostsCount={total}
+        postsCountPerPage={POSTS_COUNT_PER_PAGE}
+        pageLimit={pageLimit}
         currentPage={page}
         setCurrentPage={setCurrentPage}
       />
+
       <LoginVisible staff>
         <AdminFeatures />
       </LoginVisible>
     </PageLayout>
   );
 }
+
+const NoSearchResult = () => <p className="mx-2.5 mb-8 mt-6">검색 결과가 존재하지 않습니다.</p>;
