+++ app/[locale]/community/seminar/helper/AdminFeatures.tsx
