+++ app/[locale]/community/seminar/helper/SeminarSearchBar.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useState, ChangeEventHandler, FormEvent } from 'react';
 
 import { SearchInfo } from '@/utils/hooks/useCustomSearchParams';
