+++ app/[locale]/search/SearchPageContent.tsx
@@ -5,7 +5,7 @@ import { useTranslations } from 'next-intl';
 import { Link } from '@/navigation';
 
 import NewsRow from '@/app/[locale]/community/news/helper/NewsRow';
-import SeminarRow from '@/app/[locale]/community/seminar/SeminarRow';
+import SeminarRow from '@/app/[locale]/community/seminar/helper/SeminarRow';
 import NoticeRow from '@/app/[locale]/search/NoticeRow';
 
 import SearchBox from '@/components/common/search/SearchBox';
